% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database.R
\name{clean_database}
\alias{clean_database}
\title{Drop all tables from a database}
\usage{
clean_database(con)
}
\arguments{
\item{con}{\code{\link[DBI]{DBIConnection-class}} object}
}
\description{
Drop all tables from a database
}
\author{
John J. Aponte
}
