% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{clean_structure}
\alias{clean_structure}
\title{Clean the secondary files of the project}
\usage{
clean_structure()
}
\description{
Delete and make new \code{database}, \code{logs} and \code{reports} directory
}
\author{
John J. Aponte
}
