% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_libraries.R
\name{confirm_libraries}
\alias{confirm_libraries}
\title{Confirm if all libraries are installed}
\usage{
confirm_libraries()
}
\value{
invisible a data.frame with all the libraries found
}
\description{
Search all R and Rmd files for the keyword library(  or ::
and check if the library is installed.
}
\examples{
confirm_libraries()
}
