% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_con.R
\name{get_con}
\alias{get_con}
\title{Get conection from configuration file}
\usage{
get_con(configname = "defaultdb", file = "config.yml")
}
\arguments{
\item{configname}{a string with the name of the value}

\item{file}{name of the configuration to read from.}
}
\value{
a \code{DBI} connnection
}
\description{
The project use odbc to get connection to the database. To keep the
details of the connection we use the config.yml read by the
\code{\link[config]{get}} function.  See the vignette for details.
}
\author{
John J. Aponte
}
