% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dirs.R
\name{get_dirs}
\alias{get_dirs}
\title{Get the \code{dirs} section of the \code{config.yml} file}
\usage{
get_dirs(file = "config.yml")
}
\arguments{
\item{file}{by default the \code{config.yml} file}
}
\value{
a list with the directory entries
}
\description{
It is a wrap of \code{config::get("dirs")}.
}
