% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_con.R
\name{get_pool}
\alias{get_pool}
\title{Get a pool connection reading a configuration file}
\usage{
get_pool(configname = "defaultdb", file = "config.yml")
}
\arguments{
\item{configname}{a string with the name of the value}

\item{file}{name of the configuration to read from.}
}
\value{
a \code{\link[pool]{dbPool}} connection object
}
\description{
This function get a pool connection to a database reading the parameters from a
\code{config.yml} file using the \code{\link[config]{get}} function.
See the vignette for details.
}
\author{
John J. Aponte
}
