% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{insert_close_template}
\alias{insert_close_template}
\title{RStudio addin app to insert a template at the close part of a snip code}
\usage{
insert_close_template()
}
\value{
The template to insert
}
\description{
RStudio addin app to insert a template at the close part of a snip code
}
