% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{insert_open_template}
\alias{insert_open_template}
\title{RStudio addin app to insert a template to the heading part of a snip code}
\usage{
insert_open_template()
}
\value{
The template to insert
}
\description{
RStudio addin app to insert a template to the heading part of a snip code
}
