% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{insert_template}
\alias{insert_template}
\title{RStudio addin app to insert a template to initiate a snip code}
\usage{
insert_template()
}
\value{
The template to insert
}
\description{
RStudio addin app to insert a template to initiate a snip code
}
