% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_source.R
\name{local_source}
\alias{local_source}
\title{Run a script}
\usage{
local_source(p_script, p_log = "logs")
}
\arguments{
\item{p_script}{The script file}

\item{p_log}{the logs directory for the output}
}
\description{
Run a script in its own independent environment and logs
when it was runs, the time it takes and log the output
The file have no access to variables described outside, or
variables describe inside should not affect other programs
}
