% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database.R
\name{log_table}
\alias{log_table}
\title{Write in the log table the inclusion of a table in the database}
\usage{
log_table(con, table)
}
\arguments{
\item{con}{\code{\link[DBI]{DBIConnection-class}} object}

\item{table}{the name of the table}
}
\description{
Make a new entry in the log of the database updates.
If the log table does not exists it creates it
Make a new entry with the timestamp of the update.
In the log you will have a record of when the table was updated for the last
time. It intended to be used with \code{\link{update_table}}
}
\author{
John J. Aponte
}
