% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{master}
\alias{master}
\title{Render programs}
\usage{
master(
  start = 0,
  stop = Inf,
  format = "html",
  logdir = config::get("dirs")$logs
)
}
\arguments{
\item{start}{program to start}

\item{stop}{program to stop}

\item{format}{Format to render the programs values accepted are "pdf", "html" and "word"}

\item{logdir}{directory to keep the logs of the files. By default
the entry on the \code{config.yml} \code{dirs:logs}}
}
\value{
a data.frame with the files run, running time and exit status
}
\description{
By default, all programs with the pattern "nn_" will be executed in order.
The 'Start' and 'Stop' parameters can be used to modify the files to start or stop at
a different number.
}
\details{
The files are treated as \emph{snip} files for rmarkdown and render in the log
directory, with the format specified in the \emph{format} parameter.

The default the log directory is configured in the config.yml file.
}
