% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_snip.R
\name{master_snip}
\alias{master_snip}
\title{Render all programs as snip code}
\usage{
master_snip(
  pattern = "^[0-9][0-9].*\\\\.R$",
  start = 1,
  stop = Inf,
  logdir = config::get("dirs")$logs
)
}
\arguments{
\item{pattern}{Regular expression to select the files to run}

\item{start}{index of the program to start}

\item{stop}{index of the program to stop}

\item{logdir}{directory to keep the logs of the files. By default
the entry on the \code{config.yml} \code{dirs:logs}}
}
\value{
a data.frame with the files run, running time and exit status
}
\description{
Render the programs specified by the pattern as SPIN reports
}
