% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_spin.R
\name{master_spin}
\alias{master_spin}
\title{Render all programs as spin code}
\usage{
master_spin(
  pattern = "^[0-9][0-9].*\\\\.R$",
  start = 1,
  stop = Inf,
  logdir = config::get("dirs")$logs,
  format = "html"
)
}
\arguments{
\item{pattern}{Regular expression to select the files to run}

\item{start}{index of the program to start}

\item{stop}{index of the program to stop}

\item{logdir}{directory to keep the logs of the files. By default
the entry on the \code{config.yml} \code{dirs:logs}}

\item{format}{Format to render the report. values accepted are "pdf", "html" and "word"}
}
\value{
a data.frame with the files run, running time and exit status
}
\description{
Render the programs specified by the pattern as SPIN reports
}
