% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{A wrap to render a markdown report}
\usage{
render_report(report, format = "pdf", outputdir = get_dirs()$reports, ...)
}
\arguments{
\item{report}{filename of the report}

\item{format}{output format. "pdf","html","word" are valid entries}

\item{outputdir}{directory to save the  report}

\item{...}{other parameters for \code{\link[rmarkdown:render]{render}} function}
}
\value{
No return value, called for side effects to render the reports
}
\description{
Render the report and copy it to the \code{outputdir} directory.
More formats are available but only three are included here
`
}
\examples{
\dontrun{
render_report(myreport.rmd,"pdf")
}

}
\seealso{
\code{\link[rmarkdown]{render}}
}
