% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repana-package.R
\docType{package}
\name{repana-package}
\alias{repana}
\alias{repana-package}
\title{repana: Repeatable Analysis in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Set of utilities to facilitate the reproduction of analysis in R. It allow to make_structure(), clean_structure(), and run and log programs in a predefined order to allow secondary files, analysis and reports be constructed in an ordered and reproducible form.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaponte/repana}
  \item \url{https://johnaponte.github.io/repana/}
}

}
\author{
\strong{Maintainer}: John J. Aponte \email{john.j.aponte@gmail.com} (\href{https://orcid.org/0000-0002-3014-3673}{ORCID})

}
\keyword{internal}
