% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repana-package.R
\docType{package}
\name{repana}
\alias{repana}
\title{repana}
\description{
Reproducible Analysis in R.
}
\details{
Set of utilities to make easy the reproduction of analysis in R.
It allow to make_structure, clean_structure, and run and log programs in a
predefined order to allow secondary files, analysis and reports be constructed in
and ordered form.
}
