% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repana_targets.R
\name{targets_structure}
\alias{targets_structure}
\title{Create configuration to use targets}
\usage{
targets_structure()
}
\description{
Create of they do not exist the following directories
\itemize{
\item R
\item dat
\item out
}
}
\details{
If does not exists, create a simplify version of the \verb{_targets.R} script,
modify the \code{.gitignore}. Also add a \code{config.yml} and a \verb{_template.txt}
files.
}
\examples{
 \dontrun{
targets_structure()
}

}
