% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database.R
\name{update_table}
\alias{update_table}
\title{Helper function to include a data.frame in the database and update the log}
\usage{
update_table(con, table, source)
}
\arguments{
\item{con}{DBI connection}

\item{table}{the data.frame to be included in the database}

\item{source}{a manual comment to identify the source of the table}
}
\description{
Helper function to include a data.frame in the database and update the log
}
