% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database.R
\name{write_log}
\alias{write_log}
\title{Make a log of the updates on the database}
\usage{
write_log(con, tablename, source)
}
\arguments{
\item{con}{DBI connection}

\item{source}{a manual comment to identify the source of the table}

\item{table}{the name of the table}
}
\description{
Make a log of the database updates. If the log table does not exists it creates it
Make a new entry with the timestamp of the update
}
