% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database.R
\name{clean_database}
\alias{clean_database}
\title{Helper function to drop all tables from a database}
\usage{
clean_database(con)
}
\arguments{
\item{con}{DBI connection}
}
\description{
Helper function to drop all tables from a database
}
