% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_con.R
\name{get_con}
\alias{get_con}
\title{Get a DBI connection reading a  configuration file}
\usage{
get_con(configname = "defaultdb", file = "config.yml")
}
\arguments{
\item{configname}{a string with the name of the configuration}

\item{file}{name of the configuration to read from.}
}
\value{
a \code{DBI} connection
}
\description{
This function get a DBI connection to a database reading the parameters from a
\code{config.yml} file using the \code{\link[config]{get}} function.
See the vignette for details.
}
\author{
John J. Aponte
}
