% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_con.R
\name{get_con}
\alias{get_con}
\title{Get connection from configuration file}
\usage{
get_con(configname = "defaultdb", file = "config.yml")
}
\arguments{
\item{configname}{a string with the name of the value}

\item{file}{name of the configuration to read from.}
}
\value{
a \code{DBI} connection
}
\description{
The project use DBI to get connection to a database. To keep the
parameters of the connection we use the \code{config.yml} read by the
\code{\link[config]{get}} function.  See the vignette for details.
}
\author{
John J. Aponte
}
