% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_con.R
\name{get_con}
\alias{get_con}
\title{Get conection from configuration file}
\usage{
get_con(configname = "defaultdb", file = "config.yml", p_con)
}
\arguments{
\item{configname}{a string with the name of the value}

\item{file}{name of the configuration to read from.}

\item{p_con}{a \code{\link[DBI]{DBIConnection-class}} object.}
}
\value{
a \code{DBI} connnection
}
\description{
The project use odbc to get connection to the database. To keep the
details of the connection we use the config.yml read by the
\code{\link[config]{get}} function.  See the vignette for details.
If a \code{p_con} object is specified, the connection is passed through
}
\author{
John J. Aponte
}
