% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_source.R
\name{local_source}
\alias{local_source}
\title{Run a script}
\usage{
local_source(p_script, p_con, log = TRUE, logdir = "logs")
}
\arguments{
\item{p_script}{The script file}

\item{p_con}{a \code{\link[DBI]{DBIConnection-class}}}

\item{log}{if TRUE the log is kept}

\item{logdir}{the path to keep the log}
}
\description{
Run a script in its own independent environment and logs
when it was runs, the time it takes and log the output
The file have no access to variables described outside, or
variables describe inside should not affect other programs
except the conection to a database that is pass as parameter
}
