% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{master}
\alias{master}
\title{Runs the programs}
\usage{
master(pattern = "^[0-9][0-9].*\\\\.R$", start = 1, stop = Inf,
  logdir = config::get("dirs")$logs,
  rscript_path = "/usr/local/bin/Rscript")
}
\arguments{
\item{pattern}{Regular expression to select the files to run}

\item{start}{index of the program to start}

\item{stop}{index of the program to stop}

\item{logdir}{directory to keep the logs of the files. By default}

\item{rscript_path}{path to the \code{Rscript} file
the entry on the \code{config.yml} \code{dirs:logs}}
}
\value{
a data.frame with the files run, running time and exit status
}
\description{
Run the programs specified by the pattern, in the order as the
pattern select the files. Keep a log of the results. By default
the pattern starting with two numbers and ending with .R is selected.
They are run in order
}
\details{
The program add (or create if not exists) the files run, time of execution
and exit status in the file master.log of the directory logs.
}
