% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repana.R
\docType{package}
\name{repana}
\alias{repana}
\alias{repana-package}
\title{repana: R reproducible Analysis.}
\description{
Set of utilities to make easy the reproduction of analysis in R.
It allow to make_structure, clean_structure, and run and log programs in a
predefined order to allow secondary files, analysis and reports be constructed in
and ordered form.
}
