% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database.R
\name{update_table}
\alias{update_table}
\title{Writes a \code{data.frame} in the database}
\usage{
update_table(con, table)
}
\arguments{
\item{con}{\code{\link[DBI]{DBIConnection-class}} object}

\item{table}{Name of the table}
}
\description{
Write or replace a \code{data.frame} in the database
and update the log
}
